+++ changelog.md
@@ -5,11 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
  
-## [Release 0.1.0] - 2022-08-11
- 
+## Releases:
+
+- [Release 0.1.0] - 2022-08-11
 This is the first functional program.
- 
+
+- [Release 0.2.0] - 2022-08-30
+Separation of questions in files in the model
+
 ### Added or changed (features)
 - [TRI-1] feat(doc_project)
   - we add changelog and Readme
   - we add gitignore
+
+- [TRI-2] feat(divide_questions)
+  - We separate questions in files (model)
+  - we add more questions to test the app